requirements.txt
# Core application packages with minimal version constraints
streamlit>=1.46.0
pymongo>=4.13.0
pydantic>=2.11.0
beautifulsoup4>=4.0.0
pandas>=2.3.0
numpy>=2.3.0

# Supporting packages without version constraints
altair
annotated-types
attrs
blinker
cachetools
certifi
charset-normalizer
click
dnspython
gitdb
GitPython
idna
iniconfig
Jinja2
jsonschema
jsonschema-specifications
MarkupSafe
mccabe
mypy_extensions
narwhals
packaging
pathspec
pillow
platformdirs
pluggy
protobuf
pyarrow
pycodestyle
pydantic_core
pydeck
pyflakes
Pygments
python-dateutil
python-dotenv
pytz
referencing
requests
rpds-py
six
smmap
streamlit_iframe
tenacity
toml
tornado
typing-inspection
typing_extensions
tzdata
urllib3
watchdog
ipykernel
ipywidgets
langextract
st_pages
fuzzywuzzy
python-Levenshtein
sentence-transformers
huggingface-hub
weaviate-client